+++ doc/ChangeLog.md
@@ -9,6 +9,9 @@
 * `EDIpackSolver.superconductive_phi` has been changed to be the modulus of the
   computed superconductive order parameter. The complex argument of the order
   parameter is now accessible as `EDIpackSolver.superconductive_phi_arg`.
+* `EDIpackSolver`'s constructor got a new keyword argument `keep_dir`, which
+  disables automatic deletion of EDIpacks's temporary directory upon object
+  destruction.
 
 ## Version 0.7.0